backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
SQLAlchemy==2.0.32
pydantic==2.8.2
openai==1.40.3
pillow==10.4.0
pypdfium2==4.30.0
pandas==2.2.2
openpyxl==3.1.5
reportlab==4.2.2
anthropic==0.45.0
gunicorn==21.2.0
